Move backend config files under api/src dir
