--- config/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "lodash": "^4.17.4"
-  }
-}
